gl-iam/examples/django-keycloak/pyproject.toml
[project]
name = "django-keycloak"
version = "0.0.0"
description = "GL-IAM Django with Keycloak provider example"
requires-python = ">=3.11,<3.14"
readme = "README.md"
dependencies = [
    "django>=5.1.0,<6.0.0",
    "djangorestframework>=3.15.0,<4.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "gl-iam[django,keycloak]",
]

[tool.uv.sources]
gl-iam = { git = "https://github.com/GDP-ADMIN/gl-sdk.git", subdirectory = "libs/gl-iam", branch = "feature/gl-iam-sdk" }
